chore: rename test_evaluation_proof module to naive_evaluation_proof
