+++ crates/proof-of-sql/src/base/commitment/naive_evaluation_proof.rs
